Implemented variable module lifespans and added a base class for all command modules.
This feature allows the developers to control the instantiation behaviour for their module. There's 2 options available: singleton (module is instantiated only when created and the instance is used for all commands defined in it) and transient (module is instantiated every time a command declared in it is called, and each command gets its own instance of the module).

Furthermore, there is now a requirement for all command module classes to derive from BaseCommandModule.